refactor(types): group protocol types in module
Move key, signed packet, and resolve policy modules under
`pkarr::types` while preserving root-level re-exports.
