+++ pkarr/src/dht/mod.rs
@@ -1,0 +1,79 @@
+use ed25519_dalek::Signature;
+use mainline::MutableItem;
+use ntimestamp::Timestamp;
+
+use crate::{PublicKey, SignedPacket};
+
+impl From<&SignedPacket> for MutableItem {
+    fn from(s: &SignedPacket) -> Self {
+        Self::new_signed_unchecked(
+            s.public_key().to_bytes(),
+            s.signature().to_bytes(),
+            // Packet
+            s.inner.borrow_owner()[104..].into(),
+            s.timestamp().as_u64() as i64,
+            None,
+        )
+    }
+}
+
+impl TryFrom<&MutableItem> for SignedPacket {
+    type Error = crate::errors::SignedPacketVerifyError;
+
+    fn try_from(i: &MutableItem) -> Result<Self, crate::errors::SignedPacketVerifyError> {
+        let public_key = PublicKey::try_from(i.key())?;
+        let seq = i.seq() as u64;
+        let signature: Signature = i.signature().into();
+
+        Ok(Self {
+            inner: crate::types::signed_packet::Inner::try_from_parts(
+                &public_key,
+                &signature,
+                seq,
+                i.value(),
+            )?,
+            last_seen: Timestamp::now(),
+        })
+    }
+}
+
+impl TryFrom<MutableItem> for SignedPacket {
+    type Error = crate::errors::SignedPacketVerifyError;
+
+    fn try_from(i: MutableItem) -> Result<Self, crate::errors::SignedPacketVerifyError> {
+        SignedPacket::try_from(&i)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use mainline::MutableItem;
+
+    use crate::{Keypair, SignedPacket};
+
+    #[test]
+    fn signed_packet_converts_to_mutable_item() {
+        let keypair = Keypair::random();
+
+        let signed_packet = SignedPacket::builder()
+            .address(
+                "_derp_region.iroh.".try_into().unwrap(),
+                "1.1.1.1".parse().unwrap(),
+                30,
+            )
+            .sign(&keypair)
+            .unwrap();
+
+        let item: MutableItem = (&signed_packet).into();
+        let seq = signed_packet.timestamp().as_u64() as i64;
+
+        let expected = MutableItem::new(
+            keypair.secret_key().into(),
+            &signed_packet.packet().build_bytes_vec_compressed().unwrap(),
+            seq,
+            None,
+        );
+
+        assert_eq!(item, expected);
+    }
+}
